Split read-only MCP tool metadata
