+++ apps/main/src/server/mcp/read-only-mcp.ts
@@ -26,6 +26,15 @@ import {
   shouldShowToPublic,
 } from "@/server/db/public-visibility/filters";
 import { getClerk } from "@/server/clerk/client";
+import {
+  DEFAULT_LIMIT,
+  MAX_LIMIT,
+} from "@/server/mcp/read-only-mcp-schema-builders";
+import { buildReadOnlyMcpTools } from "@/server/mcp/read-only-mcp-tools";
+import type {
+  JsonRpcRequest,
+  McpContext,
+} from "@/server/mcp/read-only-mcp-types";
 import { searchCultivars } from "@/server/search/cultivar-search";
 
 const MCP_PROTOCOL_VERSION = "2025-11-25";
@@ -36,38 +45,6 @@ const SUPPORTED_MCP_PROTOCOL_VERSIONS = [
 ] as const;
 const MCP_SERVER_NAME = "daylily-catalog";
 const MCP_SERVER_VERSION = "0.1.0";
-const DEFAULT_LIMIT = 25;
-const MAX_LIMIT = 500;
-
-interface JsonRpcRequest {
-  jsonrpc?: string;
-  id?: string | number | null;
-  method?: string;
-  params?: unknown;
-}
-
-interface McpTool {
-  name: string;
-  title: string;
-  description: string;
-  inputSchema: Record<string, unknown>;
-  outputSchema: Record<string, unknown>;
-  securitySchemes: Array<
-    { type: "noauth" } | { type: "oauth2"; scopes: string[] }
-  >;
-  annotations?: {
-    readOnlyHint?: boolean;
-    openWorldHint?: boolean;
-    destructiveHint?: boolean;
-  };
-  _meta?: Record<string, unknown>;
-}
-
-interface McpContext {
-  baseUrl: string;
-  request: Request;
-  readDb: typeof db;
-}
 
 class McpError extends Error {
   constructor(
@@ -87,15 +64,7 @@ class McpAuthRequiredError extends Error {
 
 const REQUIRED_PRIVATE_OAUTH_SCOPE = "profile";
 const PRIVATE_OAUTH_SCOPES = getOAuthProtectedResourceScopes();
-const PUBLIC_SECURITY_SCHEMES = [{ type: "noauth" }] as const;
-const PRIVATE_SECURITY_SCHEMES = [
-  { type: "oauth2", scopes: PRIVATE_OAUTH_SCOPES },
-] as const;
-const READ_ONLY_TOOL_ANNOTATIONS = {
-  readOnlyHint: true,
-  openWorldHint: false,
-  destructiveHint: false,
-} as const;
+const tools = buildReadOnlyMcpTools(PRIVATE_OAUTH_SCOPES);
 
 function getExpectedMcpOAuthClientId() {
   const clientId = process.env.DAYLILY_MCP_OAUTH_CLIENT_ID?.trim();
@@ -194,490 +163,6 @@ const listListingsSchema = listSchema.extend({
 const getByIdSchema = z.object({
   id: z.string().trim().min(1),
 });
-
-const cultivarOutputSchema = {
-  type: "object",
-  additionalProperties: true,
-  properties: {
-    cultivar: { type: ["object", "null"], additionalProperties: true },
-  },
-};
-
-const searchResultsOutputSchema = {
-  type: "object",
-  additionalProperties: false,
-  properties: {
-    results: {
-      type: "array",
-      items: { type: "object", additionalProperties: true },
-    },
-  },
-  required: ["results"],
-};
-
-const profileOutputSchema = {
-  type: "object",
-  additionalProperties: false,
-  properties: {
-    profile: { type: ["object", "null"], additionalProperties: true },
-  },
-  required: ["profile"],
-};
-
-const paginatedOutputSchema = {
-  type: "object",
-  additionalProperties: false,
-  properties: {
-    items: {
-      type: "array",
-      items: { type: "object", additionalProperties: true },
-    },
-    nextCursor: { type: ["string", "null"] },
-  },
-  required: ["items", "nextCursor"],
-};
-
-const listOutputSchema = {
-  type: "object",
-  additionalProperties: false,
-  properties: {
-    list: { type: "object", additionalProperties: true },
-  },
-  required: ["list"],
-};
-
-const listingOutputSchema = {
-  type: "object",
-  additionalProperties: false,
-  properties: {
-    listing: { type: "object", additionalProperties: true },
-  },
-  required: ["listing"],
-};
-
-function toolMeta(invoking: string, invoked: string) {
-  return {
-    "openai/toolInvocation/invoking": invoking,
-    "openai/toolInvocation/invoked": invoked,
-  };
-}
-
-const tools: McpTool[] = [
-  {
-    name: "daylily.search_cultivars",
-    title: "Search Cultivars",
-    description:
-      "Use this when the user wants to find public daylily cultivar reference records by name, hybridizer, color, parentage, or general search text. Do not use it to answer questions about a signed-in user's own inventory; use daylily.list_listings for that. Prefer specific fields over q when the user asks for a known facet.",
-    inputSchema: {
-      type: "object",
-      additionalProperties: false,
-      properties: {
-        q: {
-          type: "string",
-          description:
-            "Broad public cultivar search text. Prefer field-specific filters when the user asks for a specific facet.",
-        },
-        cultivarName: {
-          type: "string",
-          description: "Public cultivar name filter.",
-        },
-        hybridizer: {
-          type: "string",
-          description: "Public cultivar hybridizer name filter.",
-        },
-        color: {
-          type: "string",
-          description: "Public cultivar color-notes text filter.",
-        },
-        parentage: {
-          type: "string",
-          description: "Public cultivar parentage text filter.",
-        },
-        limit: {
-          type: "integer",
-          minimum: 1,
-          maximum: MAX_LIMIT,
-          default: DEFAULT_LIMIT,
-        },
-      },
-    },
-    outputSchema: searchResultsOutputSchema,
-    securitySchemes: [...PUBLIC_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Searching cultivars", "Cultivar search complete"),
-  },
-  {
-    name: "daylily.get_cultivar",
-    title: "Get Cultivar",
-    description:
-      "Use this when the user has a Daylily Catalog cultivarReferenceId or normalizedName and needs the exact public cultivar record. Do not use it for fuzzy search; call daylily.search_cultivars first.",
-    inputSchema: {
-      type: "object",
-      additionalProperties: false,
-      properties: {
-        cultivarReferenceId: { type: "string" },
-        normalizedName: { type: "string" },
-      },
-    },
-    outputSchema: cultivarOutputSchema,
-    securitySchemes: [...PUBLIC_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading cultivar", "Cultivar loaded"),
-  },
-  {
-    name: "daylily.search_public_listings",
-    title: "Search Public Listings",
-    description:
-      "Use this when a buyer asks for public Daylily Catalog inventory, growers, prices, photos, or availability-style searches. Returns only public listings from active catalogs and never includes private notes or hidden listings. Prefer specific filters over q and follow nextCursor until null for complete result sets.",
-    inputSchema: publicListingSearchInputSchema(),
-    outputSchema: paginatedOutputSchema,
-    securitySchemes: [...PUBLIC_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Searching public listings", "Public listing search complete"),
-  },
-  {
-    name: "daylily.get_public_listing",
-    title: "Get Public Listing",
-    description:
-      "Use this when a buyer asks for one public listing by listing id, or by sellerSlug plus listingSlug. Returns public listing details only; hidden listings and private notes are excluded.",
-    inputSchema: {
-      type: "object",
-      additionalProperties: false,
-      properties: {
-        id: { type: "string", description: "Public listing id." },
-        sellerSlug: {
-          type: "string",
-          description: "Seller profile slug when using listingSlug.",
-        },
-        listingSlug: {
-          type: "string",
-          description: "Listing slug for the seller's public catalog.",
-        },
-      },
-    },
-    outputSchema: listingOutputSchema,
-    securitySchemes: [...PUBLIC_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading public listing", "Public listing loaded"),
-  },
-  {
-    name: "daylily.get_public_profile",
-    title: "Get Public Profile",
-    description:
-      "Use this when a buyer asks about a public grower or catalog profile by seller slug. Returns public seller profile, public lists, profile images, and public listing/list counts.",
-    inputSchema: publicProfileInputSchema(),
-    outputSchema: profileOutputSchema,
-    securitySchemes: [...PUBLIC_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading public profile", "Public profile loaded"),
-  },
-  {
-    name: "daylily.list_public_profile_lists",
-    title: "List Public Profile Lists",
-    description:
-      "Use this when a buyer asks what public lists or collections a seller has. Use the returned list ids with daylily.list_public_listings.",
-    inputSchema: publicProfileInputSchema(),
-    outputSchema: paginatedOutputSchema,
-    securitySchemes: [...PUBLIC_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading public lists", "Public lists loaded"),
-  },
-  {
-    name: "daylily.list_public_listings",
-    title: "List Public Listings",
-    description:
-      "Use this when a buyer asks to browse one seller's public catalog or one public list. Requires sellerSlug, and can filter by listId, listTitle, cultivar, color, price, photos, hybridizer, parentage, or year. Returns only public listings.",
-    inputSchema: publicListingSearchInputSchema({ requireSellerSlug: true }),
-    outputSchema: paginatedOutputSchema,
-    securitySchemes: [...PUBLIC_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading public listings", "Public listings loaded"),
-  },
-  {
-    name: "daylily.get_profile",
-    title: "Get Profile",
-    description:
-      "Use this when the signed-in user asks about their own Daylily Catalog profile, storefront copy, slug, logo, location, or public profile content. Requires OAuth.",
-    inputSchema: {
-      type: "object",
-      additionalProperties: false,
-      properties: {},
-    },
-    outputSchema: profileOutputSchema,
-    securitySchemes: [...PRIVATE_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading profile", "Profile loaded"),
-  },
-  {
-    name: "daylily.list_lists",
-    title: "List Lists",
-    description:
-      "Use this when the signed-in user asks to browse their catalog list records or when you need to resolve a human list name to an id before calling daylily.list_listings with listId. Returns list metadata and listing ids. Requires OAuth.",
-    inputSchema: paginatedInputSchema(),
-    outputSchema: paginatedOutputSchema,
-    securitySchemes: [...PRIVATE_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading lists", "Lists loaded"),
-  },
-  {
-    name: "daylily.get_list",
-    title: "Get List",
-    description:
-      "Use this when the signed-in user asks to inspect one of their catalog lists by id, including its description, status, and listing ids. For full listing records in a list, call daylily.list_listings with listId instead of expanding ids one by one. Requires OAuth.",
-    inputSchema: idInputSchema(),
-    outputSchema: listOutputSchema,
-    securitySchemes: [...PRIVATE_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading list", "List loaded"),
-  },
-  {
-    name: "daylily.list_listings",
-    title: "List Listings",
-    description:
-      "Use this when the signed-in user asks to search, filter, or page through their own listing records. Supports the dashboard's key advanced filters such as title, description, listId, status, price/photo availability, cultivar name, hybridizer, year, color, parentage, and bloom traits. Use the most selective filter arguments directly instead of scanning the full catalog. If the user names a list but you do not know its id, call daylily.list_lists first, then pass listId here. For complete catalog-wide summaries or facet-style searches, use limit 500 and repeat with the same filters plus nextCursor until nextCursor is null. Requires OAuth.",
-    inputSchema: listingSearchInputSchema(),
-    outputSchema: paginatedOutputSchema,
-    securitySchemes: [...PRIVATE_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading listings", "Listings loaded"),
-  },
-  {
-    name: "daylily.get_listing",
-    title: "Get Listing",
-    description:
-      "Use this when the signed-in user asks to inspect one exact listing by id, including private notes, images, linked cultivar data, and list membership. Do not use this for search or list expansion when daylily.list_listings can return the needed records in one paginated call. Requires OAuth.",
-    inputSchema: idInputSchema(),
-    outputSchema: listingOutputSchema,
-    securitySchemes: [...PRIVATE_SECURITY_SCHEMES],
-    annotations: READ_ONLY_TOOL_ANNOTATIONS,
-    _meta: toolMeta("Loading listing", "Listing loaded"),
-  },
-];
-
-tools.forEach((tool) => {
-  tool._meta = {
-    ...tool._meta,
-    securitySchemes: tool.securitySchemes,
-  };
-});
-
-function paginatedInputSchema() {
-  return {
-    type: "object",
-    additionalProperties: false,
-    properties: {
-      cursor: {
-        type: "string",
-        description:
-          "Last seen row id from nextCursor. Keep the same filters and pass this cursor to continue paging.",
-      },
-      limit: {
-        type: "integer",
-        minimum: 1,
-        maximum: MAX_LIMIT,
-        default: DEFAULT_LIMIT,
-      },
-    },
-  };
-}
-
-function listingSearchInputSchema() {
-  return {
-    type: "object",
-    additionalProperties: false,
-    properties: {
-      ...paginatedInputSchema().properties,
-      bloomHabit: {
-        type: "string",
-        description: "Bloom habit facet value, such as Diurnal or Extended.",
-      },
-      bloomSeason: {
-        type: "string",
-        description: "Bloom season facet value, such as Early, Midseason, or Late.",
-      },
-      color: {
-        type: "string",
-        description:
-          "Search linked cultivar color notes. This is better than q for requests like blue, purple eye, or green throat.",
-      },
-      cultivarName: {
-        type: "string",
-        description: "Linked cultivar name text, not listing title text.",
-      },
-      description: {
-        type: "string",
-        description: "Listing description text.",
-      },
-      foliageType: {
-        type: "string",
-        description: "Foliage type facet value.",
-      },
-      form: {
-        type: "string",
-        description: "Flower form facet value.",
-      },
-      fragrance: {
-        type: "string",
-        description: "Fragrance facet value.",
-      },
-      hasPhoto: {
-        type: "boolean",
-        description:
-          "When true, only listings with uploaded listing photos. This does not mean the linked public cultivar reference has a photo.",
-      },
-      hasPrice: {
-        type: "boolean",
-        description:
-          "When true, only listings with a positive price; when false, only listings without a positive price.",
-      },
-      hybridizer: {
-        type: "string",
-        description: "Hybridizer text from the linked cultivar record.",
-      },
-      linkedToCultivar: {
-        type: "boolean",
-        description:
-          "When true, only listings linked to a cultivar; when false, only unlinked listings.",
-      },
-      listId: {
-        type: "string",
-        description:
-          "Only listings that belong to this list id. If you only know the list name, call daylily.list_lists first.",
-      },
-      parentage: {
-        type: "string",
-        description: "Parentage text from the linked cultivar record.",
-      },
-      ploidy: {
-        type: "string",
-        description: "Ploidy facet value.",
-      },
-      priceMax: {
-        type: "number",
-        description: "Maximum listing price.",
-      },
-      priceMin: {
-        type: "number",
-        description: "Minimum listing price.",
-      },
-      q: {
-        type: "string",
-        description:
-          "Broad text search across listing title, description, private note, cultivar name, hybridizer, color, and parentage. Prefer field-specific filters when the user asks for a specific facet.",
-      },
-      status: {
-        type: "string",
-        description: "Listing status value.",
-      },
-      title: {
-        type: "string",
-        description: "Listing title text.",
-      },
-      year: {
-        type: "string",
-        description:
-          "Cultivar registration or introduction year text from the linked cultivar record, such as 2010.",
-      },
-    },
-  };
-}
-
-function publicListingSearchInputSchema(options?: { requireSellerSlug?: boolean }) {
-  return {
-    type: "object",
-    additionalProperties: false,
-    required: options?.requireSellerSlug ? ["sellerSlug"] : [],
-    properties: {
-      ...paginatedInputSchema().properties,
-      color: {
-        type: "string",
-        description: "Search linked cultivar color notes.",
-      },
-      cultivarName: {
-        type: "string",
-        description: "Linked cultivar name text.",
-      },
-      description: {
-        type: "string",
-        description: "Public listing description text.",
-      },
-      hasPhoto: {
-        type: "boolean",
-        description: "When true, only listings with uploaded listing photos.",
-      },
-      hasPrice: {
-        type: "boolean",
-        description: "When true, only listings with a positive public price.",
-      },
-      hybridizer: {
-        type: "string",
-        description: "Hybridizer text from the linked cultivar record.",
-      },
-      listId: {
-        type: "string",
-        description: "Only listings that belong to this public list id.",
-      },
-      listTitle: {
-        type: "string",
-        description: "Only listings in a public list with this title text.",
-      },
-      parentage: {
-        type: "string",
-        description: "Parentage text from the linked cultivar record.",
-      },
-      priceMax: {
-        type: "number",
-        description: "Maximum public listing price.",
-      },
-      priceMin: {
-        type: "number",
-        description: "Minimum public listing price.",
-      },
-      q: {
-        type: "string",
-        description:
-          "Broad public listing search across listing title, description, cultivar name, hybridizer, color, and parentage.",
-      },
-      sellerSlug: {
-        type: "string",
-        description: "Public seller slug or user id.",
-      },
-      title: {
-        type: "string",
-        description: "Public listing title text.",
-      },
-      year: {
-        type: "string",
-        description: "Cultivar registration or introduction year text.",
-      },
-    },
-  };
-}
-
-function publicProfileInputSchema() {
-  return {
-    type: "object",
-    additionalProperties: false,
-    required: ["sellerSlug"],
-    properties: {
-      sellerSlug: {
-        type: "string",
-        description: "Public seller profile slug or user id.",
-      },
-    },
-  };
-}
-
-function idInputSchema() {
-  return {
-    type: "object",
-    additionalProperties: false,
-    required: ["id"],
-    properties: {
-      id: { type: "string" },
-    },
-  };
-}
 
 function mcpResult(payload: unknown) {
   return {
